+++ frontend/wailsjs/go/wailsapp/App.d.ts
@@ -4,6 +4,8 @@ import {wailsapp} from '../models';
 import {state} from '../models';
 import {context} from '../models';
 
+export function ClearProjectDir():Promise<void>;
+
 export function DeleteCustomTheme(arg1:string):Promise<void>;
 
 export function DeleteWorkspace(arg1:string):Promise<void>;
@@ -17,6 +19,8 @@ export function Doctor():Promise<Array<wailsapp.CheckInfo>>;
 export function EnablePlugin(arg1:string):Promise<void>;
 
 export function GetKeybindings():Promise<Record<string, string>>;
+
+export function GetProjectDir():Promise<string>;
 
 export function GetRecent():Promise<Array<state.RecentEntry>>;
 
@@ -59,6 +63,8 @@ export function ResetKeybindings():Promise<void>;
 export function RevealPath(arg1:string):Promise<void>;
 
 export function SaveWorkspaceEditor(arg1:wailsapp.WorkspaceEditor):Promise<void>;
+
+export function SelectProjectDir():Promise<string>;
 
 export function SetCustomTheme(arg1:wailsapp.CustomThemeInfo):Promise<void>;